Decouple CSS styles from content
CSS should be generic, with classes that describe the style changes but
are not tied to the content of the page.

In this commit, I've moved as much styling as possible into generic
classes like "row" and "page" instead of using class names closely
linked to the HTML content.

Also added some better formatting of the CSS in general, with comment
headers separating related styles.