Send all data from new skill, not just name
By sending the all of the form data (using jQuery's .serialize()
function), the server can create a new skill from the form data whether
it is sent via AJAX or a regular HTTP POST request. This is useful in
case the user has disabled JavaScript.

To determine what kind of response to send back, we call `.xhr?` on the
`request` object, which returns true if the server is responding to an
AJAX request. That way, we can send back whatever the JavaScript wants
(in this case it is a snippet of HTML).