Edit button shows job edit form on résumé show page
When a user clicks the "edit" button next to a job, an edit form will
show up to allow them to edit the job information.

This is accomplished by rendering the edit forms in the page but
ensuring that they are hidden by default. This way, the JS linked to the
edit button just hides the job information and shows the form.

Currently not asynchronous.