Alter HTML structure of job listings
Makes more sense for the job info and the edit form to both be inside
the same <li> tag, since they will never be displayed at the same time
and thus act like a single list item.

This also makes it easier to associate the form with its corresponding
job info element from the JavaScript.