Implement RESTful routing conventions
RESTful routes are a way to implement a conventional and standardized
design to server routes that work with resources. In this case, the
resources we're working with are `jobs` and `skills`, so the routes that
affect those resources should follow RESTful conventions.

There are two significant changes in this commit:

1. When working with singular resources, add an identifier to the URL
   (e.g. DELETE /jobs/21 means "delete job resource with id 21)
2. Let the HTTP verb define the route (e.g. PUT /jobs/21 means
  "edit job resource with id 21, no need for an extraneous /jobs/edit in
  the route)

Check out
http://en.wikipedia.org/wiki/Representational_state_transfer#Applied_to_web_services
for more information