Remove features from /resumes/edit & rename to /users/edit
Since jobs can be edited on the main page, and skills can be created and
deleted, making them editable on a separate page seems redundant.

So instead there will just be a separate edit page for the user profile
information.

Now the app has two primary pages:

- "/" which points to "/resumes/show"
- "/users/edit"

This commit also removes the "put /skills/:skill_id" route, as it is no
longer in use anywhere.