Remove legacy code link to resumes/edit
The edit link for each job is only used by JS to show the edit job form.
As a fallback, it used to link to the /resumes/edit page. That way, in
case JS was disabled, a user would be directed to a page with an edit
form for each job that would send a (synchronous) PUT request.

Since that page no longer has an edit form for each job, it doesn't
make sense to keep this default behavior. Better to disable the link.